add teacher submodules
